backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
pydantic[email]==2.5.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

